README.md
I got 84.2% which is a HD for this assignment. 
The original file is the prototype that I made by myself without any feedback. 
The updated file is the prototype that I made after getting the feedback from my lecturer and from user-testing.
There is also a report for this assignment but I did not put it in the Github.